DevOps Ascend+ xenodata lab case (#205)
* Adding xenodata case first version

I add xenodata case study for Micorosoft Review process for case study.

* Adding AzureFunctions migration explanation and Script

I added the azure functions migration script(sync/await) and add some
explanation.

* Adding VSTS Task detail

Add an explanation for VSTS task. And pipeline for configuring k8s
cluster.

* Fix: Spell and Grammar checking

Using Grammarly I correct grammar and usage error.

* Fix: IP address masking

* Fix typo

* Update picture Adding task name

* Adding a reason why they want to move Azure.

The reason is they want to use BizSpark.

* Give some Grammar/usage check

* Clarify issue is for GCP/Current or Azure

Volker said that it is unclear which issue is for Azure or GCP. I try to
make it clarify.

* Adding an explanation why Ansible is not fit for this purpose

Volker mentioned that why Ansible is not fit for this purpose. I make it
clear and add some explanation and link for Ansible.

* Ambiguous statement. I clalify it.

I mentioned the tools which didn't work for this time. also add some link
for the product.

* Revised: fintech -> Financial serivices technology startup

* Revise the sentence into Lightweight VSM.

* Adding explanation why 30 min is long

* Fix Typo and avoid H1 by quating

* Re-arrange the format of key technologies.

* Adding explanation of why I didn't add download feature.

* Adding explanation for endpoint of VSTS

* Adding Azure Functions Code link which migrate from GCP to Azure.

* Adding typo of Customer's company name.

* Remove BizSpark price and edit it according to the Matt sugesstion

* Update 2017-01-28-xenodata.md

Copy edit in progress.

* Update 2017-01-28-xenodata.md

Copy edit in progress.

* Update 2017-01-28-xenodata.md

Copy edit in progress.

* Update 2017-01-28-xenodata.md

Copy edit in progress.

* Changing image for horizontal.

* Update 2017-01-28-xenodata.md

* Update 2017-01-28-xenodata.md