Project Origis (#116)
* Origis Template

Final description of IoT Ascend for México's project, Origis.

* Second commit

Just checking correct images path.

* Third commit

Final relative url for images.

* Four commit

All questions from Corp. first review were answered and also we found a
couple of details to correct. Now it's ready for a second review.

* Fifth commit

Problem with a couple of URLs and spelling.

* Sixth commit

Problem with a couple of URLs and spelling.

* Commit 6 - Ready for Corp. Review

Two points about security and how to handle WiFi connections are added.

* Initial pass

* Updating images

* Adding 18 and 19

* Added 18 and 19

* Updating images

* Updating image paths

* Uploading img20

* Image 20

* Fixed img20 extension

* Editing in progress

* Reorganized sections

* Edit pass #1 complete.

* Document edition

I've already worked in the remaining issues with the Origis document.

* Edit pass #2 complete

* Final edit pass

* Added {{ site.baseurl }} to images